+++ core/components/exchange/default/default-message-sender.js
@@ -84,8 +84,12 @@ class DefaultMessageSender extends MessageSender {
     disable() {
         return new Promise( ( resolve, reject ) => {
             this.isAvailable = false;
-            this.#memoryCache = null;
-            resolve();
+            this.#memoryCache.shutDown().then( () => {
+                this.#memoryCache = null;
+                resolve();
+            } ).catch( ( error ) => {
+                reject( exceptions.raise( error ) );
+            } );
         } );
     }
 }